// reiDoCangaco/src/app/models/empregado.model.ts
export interface Empregado {

    idEmpregado?: string;
    nomeEmpregado: string;
    cpf: string;
    sexo: string;
    dataNascimento: string;
    telefone: string;
    celular: string;
    email: string;
    endereco: string;
    complemento: string;
    bairro: string;
    cep: string;
    cidade: string;
    estado: string;
    pais: string;
    flagAtivo?: boolean;
    status?: string;

}